# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lunme API",
    "version": "1.0.0",
    "description": "Swagger gerado automaticamente pelas rotas do Next.js"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/internal/uploads/profile": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Faz upload de um arquivo para o storage",
        "description": "Realiza o upload de um arquivo para o bucket correto, dependendo do valor do parâmetro `path`.\n- `profile-pictures` → envia para a pasta de fotos de perfil  \n- `banner` → envia para a pasta de banners  \nO parâmetro `id` é opcional e pode ser usado como nome ou referência do arquivo.\n",
        "tags": [
          "Interno - files and uploads"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "profile-pictures",
                "banner"
              ]
            },
            "description": "Pasta de destino do arquivo",
            "example": "profile-pictures"
          },
          {
            "in": "query",
            "name": "id",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "ID opcional para identificar o arquivo",
            "example": "clzw123abc987"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Arquivo a ser enviado"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Upload realizado com sucesso!"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://supabase.../profile-pictures/abc123.png"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação (arquivo ausente ou parâmetros inválidos)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Nenhum arquivo enviado"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor ao processar o upload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro interno no servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retorna a URL pública da foto de perfil do usuário",
        "description": "Busca a imagem de perfil salva no Supabase Storage e retorna a URL pública.\nCaso o campo `photoprofile` contenha uma URL completa, apenas o path é extraído.\n",
        "tags": [
          "Interno - files and uploads"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do usuário",
            "example": "clzw0t3p8000dxoyxk2h4y1ab"
          }
        ],
        "responses": {
          "200": {
            "description": "URL pública da imagem retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string",
                      "example": "https://qfpygaqyldmthqakmisq.supabase.co/storage/v1/object/public/tiviai-images/profile-pictures/abc123.png"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de requisição (userId ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "ID do usuário não fornecido"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado ou sem foto de perfil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Usuário ou imagem não encontrada"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao gerar a URL pública",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao obter URL da imagem"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": []
}